requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data validation
pydantic==2.5.0

# Additional dependencies
python-multipart==0.0.6

# CloudBase Integration
PyMySQL
SQLAlchemy
tencentcloud-sdk-python
openai


